requirements.txt
behave==1.2.6
PyHamcrest==2.0.2
selenium==3.141.0
Appium-Python-Client==1.0.2
